Reversed the migration and fixed the unix-timestamp data type and refactored the changes to work with my code

+++ backend/src/interfaces/new_task.rs
@@ -5,7 +5,7 @@ use rocket::serde::{Deserialize, Serialize};
 pub struct NewTask {
     pub title: String,
     pub description: String,
-    pub due_date: i32, // i use UNIX-Timestamp
+    pub due_date: i64, // i use UNIX-Timestamp
     pub is_completed: bool,
     pub is_critical: bool,
 }
